Bug 544490 - Mark spy handler, menus etc as non persistent
Due to Bug 544871, the change was done in an text editor, if the model
fragment is reworked, the model fragment editor will currently remove
the tagged elements.

Also flags the generated key binding in the SpyProcessor as something
that should not get persisted. Potential follow-up, move the keybinding
creating to the model fragment instead of creating is dynamically.

Change-Id: I7e597a1bf508ac61399c72fcf8f023f0785bc9f4
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>